[compliance] restrict oscap-exec to embedded oscap-io
