+++ cmd/security-agent/subcommands/compliance/oscapexec.go
@@ -17,11 +17,10 @@ import (
 // oscapExecCommand returns the 'compliance oscap-exec' command
 func oscapExecCommand(_ *command.GlobalParams) *cobra.Command {
 	cmd := &cobra.Command{
-		Use:    "oscap-exec <binary-path> [args...]",
+		Use:    "oscap-exec [args...]",
 		Short:  "Execute oscap-io with dropped capabilities (internal use only)",
 		Long:   "Internal command that drops capabilities before executing oscap-io. This command should not be called directly by users.",
 		Hidden: true,
-		Args:   cobra.MinimumNArgs(1),
 		RunE: func(_ *cobra.Command, args []string) error {
 			return cli.RunOscapExec(args)
 		},
